Switch primary keys from TSID to UUIDv7
The TSID node field gives ~38 concurrent generators before birthday
collisions become likely, and the only mitigation was manual
coordination through RatchetOptions.tsidNodeId. UUIDv7 has 62 bits of
randomness in rand_b, so coordination falls away, and the millisecond
timestamp prefix preserves the same time-ordered B-tree locality TSIDs
were chosen for in the first place.

Pre-1.0 break, no compatibility shim. java.util.UUID flows through
every persistent entity, the public JobSchedulerService API,
JobHandle, JobContext, and all event payloads. cancelJob, pauseJob,
resumeJob, retryJob, and replace all take UUID jobId. JobHandle.id()
returns UUID. JobContext.jobId is UUID. AbstractJobSchedulerEvent
and its 16 subclasses carry UUID jobId. MetricsCollector and
ResultPersistenceStrategy SPIs flip in lockstep.

UuidV7Factory implements RFC 9562 §5.7. The 12-bit rand_a is used as
a per-millisecond monotonic counter; on overflow within one ms it
busy-spins via Thread.onSpinWait until the wall clock advances. That
is the §6.2 wait-for-tick variant. The timestamp-increment
alternative was rejected because advancing the embedded timestamp
ahead of wall-clock time would corrupt time-correlation queries. No
new runtime dependencies — only java.security.SecureRandom and
java.util.UUID.

Storage is per-store native:

  PostgreSQL: native uuid (16 bytes). Hibernate's PostgreSQLDialect
  maps java.util.UUID automatically. No entity-side annotations.

  MySQL: BINARY(16), again via Hibernate's default mapping. A
  read-only vw_jobs view ships in ddl/views/, wrapping every UUID
  column in BIN_TO_UUID(col, 1) so operators can read hyphenated
  UUID strings directly. The "1" matters: without it MySQL applies
  a timestamp-shift transform that does not round-trip with Java's
  UUID.toString().

  MongoDB: BSON binary subtype 4 via UuidRepresentation.STANDARD.
  MongoClientFactory enforces this on caller-supplied clients, and a
  @PostConstruct probe on MongoJobStoreImpl encodes a known UUID at
  startup and inspects the subtype byte. Anything other than 4
  throws RatchetConfigurationException. Silent UUID corruption from
  JAVA_LEGACY byte-order swaps is not detectable post-rollout, so
  the validator is fail-fast.

The TCK schema-conformance contract gains LogicalType.UUID. The
catalog enumerates every PK and FK across all 15 scheduler tables
and asserts UUID for them; per-store dialect mappers translate to
native uuid and BINARY. MySQL DDL also gained the fk_bk_owner_job
constraint on scheduler_business_key_reservation that PostgreSQL has
carried since v3, so both stores now enforce CASCADE-on-job-delete
for active business key reservations.

RatchetOptions.NodeBuilder.tsidNodeId(int) and the
RATCHET_TSID_NODE_ID config key are removed. Other NodeOptions
fields (heartbeatIntervalSeconds, orphanGraceSeconds, and friends)
are unchanged.

At scale (100M rows × 5 secondary indexes), the 16-byte UUID PK
adds roughly 4.8 GB of index space versus the prior 8-byte BIGINT
design. Documented as a known characteristic in ARCHITECTURE.md and
the Q7 ADR.

See docs/decisions/Q7-uuidv7-pk.md for the full architectural
rationale and the spike outcomes that confirmed each storage choice.

+++ ratchet-store-mongodb/src/test/java/run/ratchet/store/mongodb/WorkflowConditionIT.java
@@ -51,7 +51,7 @@ class WorkflowConditionIT extends BaseDocumentStoreIT {
 
     List<WorkflowConditionEntity> found = store().findConditionsByChildJobId(child.getId());
     assertEquals(1, found.size());
-    assertEquals(parent.getId(), (long) found.get(0).getParentJobId());
+    assertEquals(parent.getId(), found.get(0).getParentJobId());
   }
 
   @Test
